Revision of the homework
1- Change the float to double
2- Fixing the issues after the changes
3- Fixing the main code

+++ src/vehicle/Drivable.java
@@ -9,11 +9,11 @@ package vehicle;
  * @author sweis
  */
 public interface Drivable {
-    public void accelerate(float speed);
+    public void accelerate(double speed);
     public void brake();
-    public void turn(float angle);
-    public float getDirection(); 
-    public float getSpeed();   
+    public void turn(double angle);
+    public double getDirection(); 
+    public double getSpeed();   
     public String getMake();    
     public String getType();
 }
