+++ src/vehicle/Flyable.java
@@ -9,6 +9,6 @@ package vehicle;
  * @author sweis
  */
 public interface Flyable extends Drivable {
-    public void changeAltitude(float change);
-    public float getAltitude();
+    public void changeAltitude(double change);
+    public double getAltitude();
 }
